+++ fabric/src/main/java/net/dawson/adorablehamsterpets/fabric/client/AdorableHamsterPetsFabricClient.java
@@ -34,6 +34,7 @@ public final class AdorableHamsterPetsFabricClient implements ClientModInitializ
         KeyMappingRegistry.register(ModKeyBindings.GENETICS_VISUALIZER_VAR_DOWN_KEY);
         KeyMappingRegistry.register(ModKeyBindings.GENETICS_VISUALIZER_MUT_UP_KEY);
         KeyMappingRegistry.register(ModKeyBindings.GENETICS_VISUALIZER_MUT_DOWN_KEY);
+        KeyMappingRegistry.register(ModKeyBindings.TOGGLE_PERFORMANCE_MODE_KEY);
 
         // --- Register Particle Provider ---
         for (RegistrySupplier<SimpleParticleType> particleSupplier : ModParticles.BEDDING_PARTICLES.values()) {